Do not crash if an object doesn't have a name
Sadly, there are BACnet devices in the wild that violate
ISO 16484-5-2014 H.5.4.9 "Defining Proprietary Object Types" and provide
Proprietary Objects that are missing the Object_Name property. One
example is the Delta Controls enteliTouch eTCH-7ET.

Yabe currently panics and crashes if the device returns an "unknown
property" error while listing object names. This patch makes it log a
warning and carry on instead of crashing.