Be more lenient when parsing Time APDUs
ISO 16484-5-2014 20.2.13 "Encoding of a Time Value" states:

> A value of X'FF' = D'255' in any of the four octets shall indicate
> that the corresponding value is unspecified and shall be considered a
> wildcard when matching times. If all four octets = X'FF', the
> corresponding time may be interpreted as "any" or "don't care."

Yabe doesn't properly handle the case where some of the octets are set
to 0xFF but not others, which is a valid case according to the above.

This can cause actual problems in practice, for example it makes it
impossible to list some object properties from Delta Controls DAC-1180E
controllers which sometimes return Time APDUs with hour = min = sec =
0xFF but hundreds = 0. This in turn makes it impossible to export the
Device DB.

This patch fixes the issue by making Yabe more tolerant of unusual Time
APDUs.